Removed and added new console logs
Added a new error to check for when the password required fails.  If it does, we check what the password is set to.  If it set to our access code, then we refresh the page so that we can use the proper authentication methods (which will not hit this error).  If it is a user set password, then we allow to error to proceed.